Experimenting
Added two additional platforms in the CSG testing scene - both are for testing which texture of grass looks better, and also for deciding whether I should use procedural noise to generate textures to optimize the size of the game.

Yes, I do focus on details like this. Yes, I know it's stupid.