ehhhh screw it
getting entities to work (PROPERLY) under Trenchbroom is going to be a pain in the ass, so........instead I've decided that buttons will be implemented in the Godot editor.

yes, I know that this is going to be awful for people who would try forking this project and making their own maps, but I really can't think of any other effective way to implement an interaction system.

now that i think of it, i could try checking out Qodot's (func_godot's predecessor's) repository for some examples that could help. func_godot doesn't have any, unfortunately.